[Fix] global 패키지 생성
Context 및 Filter 관련 클래스를 하단에 위치
