+++ src/main/java/codesquad/ui/UserRegisterFormServlet.java
@@ -1,4 +1,4 @@
-package codesquad.servlet;
+package codesquad.ui;
 
 import jakarta.servlet.ServletConfig;
 import jakarta.servlet.ServletException;
@@ -11,22 +11,21 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 
-@WebServlet(urlPatterns = "/logout")
-public class LogoutServlet extends HttpServlet {
-    private static final Logger logger = LoggerFactory.getLogger(LogoutServlet.class);
+@WebServlet(urlPatterns = "/users/register-form")
+public class UserRegisterFormServlet extends HttpServlet {
+    private static final Logger logger = LoggerFactory.getLogger(UserRegisterFormServlet.class);
 
     @Override
     public void init(ServletConfig config) throws ServletException {
         super.init(config);
-        logger.info("LogoutServlet initialized");
+        logger.info("UserRegisterFormServlet initialized");
     }
 
     /**
-     * 로그아웃 요청
+     * 유저 등록 폼 요청
      */
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getSession().invalidate();
-        resp.sendRedirect("/");
+        req.getRequestDispatcher("/WEB-INF/views/user/form.jsp").forward(req, resp);
     }
 }
